requirements.txt
PySide6>=6.6
Pillow>=10.0
numpy>=1.24
python-dotenv>=1.0
anthropic>=0.34
manga-ocr>=0.1.11
easyocr>=1.7
# Torch is pulled in by manga-ocr / easyocr. Install a CUDA build manually
# beforehand if you want GPU acceleration.
